+++ project/src/api/drop-token/drop-token.ts
@@ -1,0 +1,7 @@
+import {AUTH_TOKEN_KEY_NAME} from '../../const';
+
+const dropToken = (): void => {
+  localStorage.removeItem(AUTH_TOKEN_KEY_NAME);
+};
+
+export default dropToken;
