+++ project/src/api/index.ts
@@ -1,0 +1,4 @@
+export {default as getToken} from './get-token/get-token';
+export {default as saveToken} from './save-token/save-token';
+export {default as dropToken} from './drop-token/drop-token';
+export {default as createAPI} from './create-api/create-api';
