+++ project/src/const/auth-token-key-name/auth-token-key-name.ts
@@ -1,0 +1,3 @@
+const AUTH_TOKEN_KEY_NAME = 'six-cities-token';
+
+export default AUTH_TOKEN_KEY_NAME;
