+++ project/src/const/index.ts
@@ -6,3 +6,4 @@ export {default as ReviewRate} from './review-rate/review-rate';
 export {default as Cities} from './cities/cities';
 export {default as OffersSortingType} from './offers-sorting-type/offers-sorting-type';
 export {default as AppStatus} from './app-status/app-status';
+export {default as AUTH_TOKEN_KEY_NAME} from './auth-token-key-name/auth-token-key-name';
