+++ project/src/store/index.ts
@@ -1,5 +1,5 @@
 import {configureStore} from '@reduxjs/toolkit';
-import {createAPI} from '../api/api';
+import {createAPI} from '../api';
 import {reducer} from './reducer';
 
 const api = createAPI();
