+++ project/src/types/token/token.ts
@@ -1,0 +1,3 @@
+type Token = string;
+
+export default Token;
